LayoutTests/imported/w3c/web-platform-tests/content-security-policy/style-src/style-src-inline-style-blocked-expected.txt


PASS Inline style element should not load without 'unsafe-inline'
FAIL Should fire a securitypolicyviolation event assert_equals: expected "style-src" but got "style-src-elem"
Lorem ipsum
